pyproject.toml
# pyproject.toml
[project]
name = "mcp-chakoshi"
version = "0.1.0"
description = "A minimal MCP server that bridges Claude Desktop and chakoshi moderation API"
requires-python = ">=3.10"

dependencies = [
  "mcp[cli]>=1.6.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1"
]

[project.scripts]
mcp-chakoshi = "chakoshi_server.server:run_stdio"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chakoshi_server"]
